#83/1-consumer/requirements.txt
python-dotenv==0.16.0
pandas==1.2.3
numpy==1.20.2
confluent-kafka==1.9.2
avro-python3==1.10.2
jsonschema==3.2.0
lensesio==4.0.1

